+++ src/main/java/com/gotcha/server/auth/controller/TokenController.java
@@ -1,11 +1,13 @@
 package com.gotcha.server.auth.controller;
 
+import com.gotcha.server.auth.dto.request.MemberDetails;
 import com.gotcha.server.auth.dto.response.RefreshTokenResponse;
 import com.gotcha.server.auth.service.TokenService;
 import com.gotcha.server.member.dto.request.RefreshTokenRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,5 +22,12 @@ public class TokenController {
     public ResponseEntity<RefreshTokenResponse> refreshToken(
             @RequestBody final RefreshTokenRequest request) {
         return ResponseEntity.ok(tokenService.refresh(request));
+    }
+
+    @PostMapping("/api/logout")
+    @Operation(description = "로그아웃한다.")
+    public ResponseEntity<Void> logout(@AuthenticationPrincipal final MemberDetails details) {
+        tokenService.removeRefreshToken(details);
+        return ResponseEntity.ok().build();
     }
 }
