+++ src/main/java/com/gotcha/server/auth/repository/RefreshTokenRepository.java
@@ -6,4 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findBySocialId(String socialId);
+    void deleteBySocialId(String socialId);
 }
